feat: update id type from int to long

+++ Mecanillama.API/Security/Domain/Services/IUserService.cs
@@ -1,4 +1,5 @@
 ﻿using Mecanillama.API.Security.Domain.Models;
+using Mecanillama.API.Security.Domain.Services.Communication;
 
 namespace Mecanillama.API.Security.Domain.Services;
 
@@ -6,8 +7,8 @@ public interface IUserService
 {
     Task<AuthenticateResponse> Authenticate(AuthenticateRequest request);
     Task<IEnumerable<User>> ListAsync();
-    Task<User> GetByIdAsync(int id);
+    Task<User> GetByIdAsync(long id);
     Task RegisterAsync(RegisterRequest request);
-    Task UpdateAsync(int id, UpdateRequest request);
-    Task DeleteAsync(int id);
+    Task UpdateAsync(long id, UpdateRequest request);
+    Task DeleteAsync(long id);
 }
